+++ Changes.md
@@ -9,10 +9,10 @@ developed by
 
 ## New 
 
-* NEW (13.01.2024) Added basic 'http' and 'json' import libraries.
+* NEW (20.01.2024) Added JavaFX classes to Dpdl class definition file and created Dpdl example app that uses JavaFX to render UI elements via FXML
 
-* NEW (02.23.2023) JavaScript can now be executed also via the 'QuickJS' javascript engine (ES2020 spec) by using the keyword '>>qjs'
-
+* (13.01.2024) Added basic 'http' and 'json' import libraries.
+* (02.23.2023) JavaScript can now be executed also via the 'QuickJS' javascript engine (ES2020 spec) by using the keyword '>>qjs'
 * (27.12.2023) 'import' statement for libraries
 * (19.12.2023) 	Function return type definitions added. Type 'enum' is supported. Function 'convert(..)' added
 * (14.12.2023) 'for' statement is available i a first draft, it will be worked out further